+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Implemented GET /api/v1/@apostrophecms/login/whoami route such that it returns the details of the currently logged in user; added the route to the login module.
   Thanks to [sombitganguly](https://github.com/sombitganguly) for this contribution.
 * Adds keyboard shortcuts for manipulating widgets in areas. Includes Cut, Copy, Paste, Delete, and Duplicate.
-* Adds dynamic choices working with piece manager filters. 
+* Adds dynamic choices working with piece manager filters.
 * Allow `import.imageTags` (array of image tag IDs) to be passed to the rich text widget when importing (see https://docs.apostrophecms.org/reference/api/rich-text.html#importing-inline-images).
 
 ### Changes
@@ -16,6 +16,7 @@
 
 * Add missing Pages manager shortcuts list helper.
 * Improve the `isEmpty` method of the rich text widget to take into account the HTML blocks (`<figure>` and `<table>`) that are not empty but do not contain any plain text.
+* Fixed admin bar item ordering to correctly respect the precedence hierarchy: groups (when leader is positioned) > explicit order array > groups (when leader has positioning options) > individual `last`/`after` options.
 * (Backward compatibility break) Conditional field that depends on already hidden field is also hidden, again.
 
 ## 4.18.0 (2025-06-11)